refactor: move blueprint import and save paths into blueprint_write
tools_blueprints.py sat at 997 lines, the styleguide's module-size review
signal. Keep it as the tool surface (dispatch, list/get/delete/substitute)
and move the two blueprint/save-backed write paths, with the WebSocket error
helpers they share, into tools/blueprint_write.py. Behaviour and the unit
tests are unchanged.
